requirements.txt
aiogram
python-dotenv
pillow
loguru
